update : article / comments user 인가 기능 추가

+++ controllers/article-controllers.js
@@ -1,14 +1,29 @@
 import prisma from '../lib/prismaclient.js';
 
-export async function articleNew(req, res) {
+export async function createArticle(req, res) {
+  // user가 DB에 존재 하는지 확인
+  const userId = req.user.id;
+  const findUser = await prisma.user.findUnique({ where: { id: userId } });
+
+  console.log(findUser);
+
+  if (!findUser) return res.status(401).json({ message: 'Unauthorized' });
+
+  // article 저장하기
+  const { title, content } = req.body;
+
   const articleCreate = await prisma.article.create({
-    data: req.body,
+    data: {
+      title,
+      content,
+      userId,
+    },
   });
 
   res.status(201).json(articleCreate);
 }
 
-export async function articlesList(req, res) {
+export async function getArticlesList(req, res) {
   const {
     offset = 0,
     limit = 5,
@@ -55,7 +70,7 @@ export async function articlesList(req, res) {
   res.status(200).json(articles);
 }
 
-export async function articleOnly(req, res) {
+export async function getArticleInfo(req, res) {
   const id = req.params.id;
   const article = await prisma.article.findUniqueOrThrow({
     where: { id },
@@ -72,33 +87,55 @@ export async function articleOnly(req, res) {
   res.status(200).json(article);
 }
 
-export async function articleUpdate(req, res) {
-  const id = req.params.id;
+export async function updateArticle(req, res) {
+  const articleId = req.params.id;
+  const userId = req.user.id;
 
-  const article = await prisma.article.findUnique({ where: { id } });
+  // article이 DB에 있는지 확인
+  const article = await prisma.article.findUnique({ where: { id: articleId } });
 
   if (!article)
     return res.status(401).json({ message: 'Cannot found article' });
 
+  // User가 DB에 존재 하는지 확인
+  const findUser = await prisma.user.findUnique({ where: { id: userId } });
+
+  if (!findUser) return res.status(401).json({ message: 'Unauthorized' });
+
+  // DB에 있는 article userID 정보와 로그인 한 User 정보가 같은지 확인
+  if (article.userId !== userId)
+    return res.status(401).json({ message: 'Unauthorized' });
+
   const articleUpdate = await prisma.article.update({
-    where: { id },
+    where: { id: articleId },
     data: req.body,
   });
 
   res.status(200).json(articleUpdate);
 }
 
-export async function articleDelete(req, res) {
-  const id = req.params.id;
+export async function deleteArticle(req, res) {
+  const articleId = req.params.id;
+  const userId = req.user.id;
 
-  const article = await prisma.product.findUnique({ where: { id } });
+  // Article이 DB에 있는지 확인
+  const article = await prisma.article.findUnique({ where: { id: articleId } });
 
   if (!article)
     return res.status(401).json({ message: 'Cannot found article' });
 
+  // User가 DB에 존재 하는지 확인
+  const findUser = await prisma.user.findUnique({ where: { id: userId } });
+
+  if (!findUser) return res.status(401).json({ message: 'Unauthorized' });
+
+  // DB에 있는 article userID 정보와 로그인 한 User 정보가 같은지 확인
+  if (article.userId !== userId)
+    return res.status(401).json({ message: 'Unauthorized' });
+
   await prisma.article.delete({
-    where: { id },
+    where: { id: articleId },
   });
 
-  res.status(204).json({ massage: `delete Article ${id}` });
+  res.status(204);
 }
